pocketflow-leetcode/leetcode-agent/requirements.txt
# Core dependencies
pocketflow>=0.1.0

# LLM and API
openai>=1.0.0
requests>=2.28.0

# Web scraping
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Data validation
pydantic>=1.10.0
PyYAML>=6.0

# Code execution
execnet>=1.9.0

# Utilities
python-dotenv>=0.19.0
rich>=12.0.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=3.0.0